Fixed the Client
The Processing Client is working now.
My problem was, that the header needs to be a 32 bit integer, built from
bytes.
The example is sending a "Ping" from BWS to the Client which is sending
it back to BWS where it is getting re-constructed into a String.
The Server will be fixed tomorrow ;-)